refactor(hmsl): move show_results function
