+++ ggshield/cmd/hmsl/check.py
@@ -5,7 +5,6 @@ import click
 from requests import HTTPError
 
 from ggshield.cmd.common_options import add_common_options, json_option
-from ggshield.cmd.hmsl.decrypt import show_results
 from ggshield.cmd.hmsl.hmsl_common_options import (
     full_hashes_option,
     input_arg,
@@ -17,6 +16,7 @@ from ggshield.core.errors import UnexpectedError
 from ggshield.core.text_utils import display_info, pluralize
 from ggshield.hmsl import Secret, get_client
 from ggshield.hmsl.collection import InputType, NamingStrategy, collect, prepare
+from ggshield.hmsl.output import show_results
 
 
 logger = logging.getLogger(__name__)
