+++ ggshield/cmd/hmsl/decrypt.py
@@ -1,25 +1,17 @@
 import json
-from typing import Any, Dict, Iterable, Iterator, TextIO, cast
+from typing import Any, Dict, Iterator, TextIO, cast
 
 import click
 
 from ggshield.cmd.common_options import add_common_options, json_option
 from ggshield.cmd.hmsl.hmsl_common_options import input_arg
 from ggshield.core.errors import ParseError
-from ggshield.core.text_utils import display_heading, display_warning, pluralize
 from ggshield.hmsl import Match, Secret
 from ggshield.hmsl.crypto import make_hint
+from ggshield.hmsl.output import show_results
 
 
 # Types and constants
-TEMPLATE = """
-> Secret {number}
-Secret name: "{name}"
-Secret hash: "{hash}"
-Distinct locations: {count}
-First occurrence:
-    URL: "{url}"
-"""
 
 
 # Command
@@ -93,36 +85,3 @@ def decrypt(input: TextIO, mapping: Dict[str, str]) -> Iterator[Secret]:
                 yield match.decrypt(hints[match.hint])
         else:
             yield Secret(**data)
-
-
-def show_results(
-    secrets: Iterable[Secret], names: Dict[str, str], json_output: bool
-) -> None:
-    """
-    Display the secrets.
-    """
-    secrets = list(secrets)
-    if secrets:
-        display_warning(
-            f"Found {len(secrets)} leaked {pluralize('secret', len(secrets))}."
-        )
-    else:
-        display_heading("All right! No leaked secret has been found.")
-
-    data = {
-        "leaks_count": len(secrets),
-        "leaks": [
-            {
-                "name": names.get(secret.hash) or secret.hash,
-                "hash": secret.hash,
-                "count": secret.count,
-                "url": secret.url,
-            }
-            for secret in secrets
-        ],
-    }
-    if json_output:
-        click.echo(json.dumps(data))
-    else:
-        for i, secret in enumerate(data["leaks"]):
-            click.echo(TEMPLATE.format(number=i + 1, **secret))
